requirements.txt
discord.py >= 1.7.3, <= 1.8.0
pytest >= 6.2.5, <= 6.3.0
pytest-dotenv >= 0.5.2, <= 0.6.0
pylint >= 2.11.1, <= 2.12.0
python-dotenv~=1.0.0
setuptools~=58.1.0